+++ scripts/homebrew-update.mjs
@@ -52,7 +52,7 @@ const formula = `class ${formulaClass} < Formula
 
   def install
     ENV["HOME"] = buildpath
-    system "npm", "install", *std_npm_install_args
+    system "npm", "install", *std_npm_args
   end
 
   test do